// src/app/models/gif.response.ts
export interface Images {
  fixed_width_still: FixedWidthStill;
  preview_gif: PreviewGif;
  fixed_height_downsampled: FixedHeightDownsampled;
  preview: Preview;
  fixed_height_small: FixedHeightSmall;
  downsized: Downsized;
  fixed_width_downsampled: FixedWidthDownsampled;
  fixed_width: FixedWidth;
  downsized_still: DownsizedStill;
  downsized_medium: DownsizedMedium;
  original_mp4: OriginalMp4;
  downsized_large: DownsizedLarge;
  preview_webp: PreviewWebp;
  original: Original;
  original_still: OriginalStill;
  fixed_height_small_still: FixedHeightSmallStill;
  fixed_width_small: FixedWidthSmall;
  looping: Looping;
  downsized_small: DownsizedSmall;
  fixed_width_small_still: FixedWidthSmallStill;
  fixed_height_still: FixedHeightStill;
  fixed_height: FixedHeight;
}

export interface User {
  avatar_url: string;
  banner_image: string;
  banner_url: string;
  profile_url: string;
  username: string;
  display_name: string;
  description: string;
  is_verified: boolean;
  website_url: string;
  instagram_url: string;
}

export interface Data {
  type: string;
  id: string;
  url: string;
  slug: string;
  bitly_gif_url: string;
  bitly_url: string;
  embed_url: string;
  username: string;
  source: string;
  title: string;
  rating: string;
  content_url: string;
  source_tld: string;
  source_post_url: string;
  is_sticker: number;
  import_datetime: string;
  trending_datetime: string;
  images: Images;
  user: User;
}

export interface Meta {
  msg: string;
  status: number;
  response_id: string;
}

export interface GifResponse {
  data: Data;
  meta: Meta;
}